refactor: 구간 역 제거로직 Line으로 이동, 단위테스트 작성
